chore: remove need for extract_alerts.sh
Make excels at tracking generated output files from specified input
files (it might be the only thing it's good at).

By having a convention of naming the generated rule files and their
corresponding unit test files based on the file names in the
ConfigMap, then we don't need a separate script to track the
difference in what the files are called inside and outside the
ConfigMap, which would need to be updated each time a new unit test
file is added. With this change, as long as you name your unit test
file in the appropriate way, and there's a corresponding entry in the
ConfigMap, then the rule file should get extracted in the expected
way.